Extract promotion logic from state.py
